split into SetFormulaAlg
